requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bitmath==1.3.3.1
cachetools==5.5.2
certifi==2025.8.3
chardet==4.0.0
charset-normalizer==3.4.3
cheesechaser==0.2.2
click==8.2.1
deprecation==2.1.0
distro==1.9.0
docstring-parser==0.17.0
fastapi==0.116.1
filelock==3.19.1
fsspec==2025.9.0
ftfy==6.3.1
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-aiplatform==1.112.0
google-cloud-bigquery==3.37.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.36.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.74.0
grpcio-status==1.74.0
h11==0.16.0
h2==4.3.0
hbutils==0.11.0
hf-xet==1.1.9
hfutils==0.11.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
jinja2==3.1.6
jiter==0.10.0
markupsafe==3.0.2
mpmath==1.3.0
natsort==8.4.0
networkx==3.5
numpy<2.3.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
open-clip-torch==3.1.0
openai==1.107.1
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.2
pillow==11.3.0
portalocker==3.2.0
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic==2.11.7
pydantic-core==2.33.2
pyrate-limiter==3.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytimeparse==1.1.8
pytz==2025.2
pyyaml==6.0.2
qdrant-client==1.15.1
random-user-agent==1.0.1
regex==2025.9.1
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
setuptools==80.9.0
shapely==2.1.1
six==1.17.0
sniffio==1.3.1
starlette==0.47.3
sympy==1.14.0
tenacity==9.1.2
timm==1.0.19
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
triton==3.4.0
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
urlobject==3.0.0
uvicorn==0.35.0
wcwidth==0.2.13
websockets==15.0.1
